rename CatalogTableAsV2 to UnresolvedTable
