--- sql/catalyst/src/main/java/org/apache/spark/sql/sources/v2/internal/UnresolvedTable.java
@@ -1,28 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.sources.v2.internal;
-
-import org.apache.spark.sql.sources.v2.Table;
-
-/**
- * Internal interface used for table definitions, which we do not have complete information to
- * resolve yet. This is primarily used by the `CatalogTableAsV2` wrapper returned by the
- * V2SessionCatalog. When a `CatalogTableAsV2` is returned by the V2SessionCatalog, we defer
- * planning to V1 data source code paths.
- */
-public interface UnresolvedTable extends Table {}
